requirements.txt
googletrans==4.0.0-rc1
gTTS==2.5.1
facebook-sdk==3.1.0
google-api-python-client==2.149.0
youtube-transcript-api==0.6.2
requests==2.31.0
